requirements.txt
requests
beautifulsoup4
tqdm
Flask
python-dotenv
Flask-Cors
google-generativeai
gunicorn